Move color dictionaries to submodule and don't evaluate coverage
Update linter exclusions

Fix color dictionary imports in palette tests
