+++ Makefile
@@ -10,4 +10,4 @@ coverage:
 	pytest --doctest-modules --cov=seaborn --cov-config=.coveragerc seaborn
 
 lint:
-	flake8 --exclude seaborn/__init__.py,seaborn/cm.py,seaborn/tests,seaborn/external seaborn
+	flake8 --exclude seaborn/__init__.py,seaborn/colors/__init__.py,seaborn/cm.py,seaborn/tests,seaborn/external seaborn
