+++ seaborn/colors/__init__.py
@@ -1,0 +1,2 @@
+from .xkcd_rgb import xkcd_rgb
+from .crayons import crayons
