+++ seaborn/colors/crayons.py
